requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
PyJWT==2.8.0
requests==2.31.0
urllib3==2.2.0